+++ loan.html
@@ -92,7 +92,7 @@
         </div>
     </div>
 
-    <script src="js/loan_calculator.js"></script>
+    <script src="js/main_loan_calculator.js"></script>
 </body>
 
 </html>
